requirements.txt
boto3==1.7.33
botocore==1.10.33
docutils==0.14
jmespath==0.9.3
python-dateutil==2.7.3
PyYAML==3.12
s3transfer==0.1.13
six==1.11.0
